Reorganized tests directory
