+++ tests/deployment/test_language_container_deployer_cli.py
@@ -4,7 +4,7 @@ from click.testing import CliRunner
 from exasol_advanced_analytics_framework import deploy
 from tests.utils.revert_language_settings import revert_language_settings
 from tests.utils.db_queries import DBQueries
-from tests.utils.parameters import bucketfs_params, db_params
+from tests.utils import bucketfs_params, db_params
 from pathlib import Path
