+++ tests/test_package/pyproject.toml
@@ -1,0 +1,16 @@
+[tool.poetry]
+name = "test_event_handlers"
+version = "0.1.0"
+description = ""
+authors = ["Umit Buyuksahin <umit.buyuksahin@exasol.com>"]
+
+include = ["../test_event_handlers/*"]
+
+[tool.poetry.dependencies]
+python = "^3.8"
+
+[tool.poetry.dev-dependencies]
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
